+++ test/resources/example-data.js
@@ -47,6 +47,13 @@ const EX_4_HTML_CONVERTER_OUTPUT = '<!DOCTYPE html><html lang="en"><head><title>
   '<p>This is an introduction to something cool.</p>\n' +
   '</body></html>';
 
+const EX_5_HTML_CONVERTER_HTML_INFO = {
+  engine: 'md',
+  file: path.resolve('test/resources/layout-in-layout-test/page.md'),
+  layoutsDir: path.resolve('test/resources/layout-in-layout-test/')
+};
+const EX_5_HTML_CONVERTER_OUTPUT = '<body><p>test</p>\n</body>';
+
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 //                                                                                                                    //
 //                                                    RdfConverter                                                    //
@@ -89,6 +96,8 @@ module.exports = {
   EX_3_HTML_CONVERTER_OUTPUT,
   EX_4_HTML_CONVERTER_HTML_INFO,
   EX_4_HTML_CONVERTER_OUTPUT,
+  EX_5_HTML_CONVERTER_HTML_INFO,
+  EX_5_HTML_CONVERTER_OUTPUT,
   EX_1_RDF_CONVERTER_GRAPHQLLD,
   EX_1_RDF_CONVERTER_DATA,
 };